fix(migrations)!: add list foreign key constraint actions
this is a "db breaking change"! it is not really noticeable mostly, but
will prevent deleting objects/actors cleanly if present in a list. this
updates the migration so that new dbs won't be affected, but existing
dbs should manually update lists and list_elements tables to have
CASCADE on delete actions for all foreign keys!